+++ cmd/web/main.go
@@ -1,0 +1,25 @@
+package main
+
+import (
+	"log"
+	"net/http"
+)
+
+func main() {
+
+	// Register the two new handler functions and corresponding URL patterns with // the servemux, in exactly the same way that we did before.
+
+	mux := http.NewServeMux()
+	mux.HandleFunc("/", home)
+	mux.HandleFunc("/snippet/view", snippetView)
+	mux.HandleFunc("/snippet/create", snippetCreate)
+
+	fileServer := http.FileServer(http.Dir("./ui/static/"))
+	// Use the mux.Handle() function to register the file server as the handler for     // all URL paths that start with "/static/". For matching paths, we strip the     // "/static" prefix before the request reaches the file server.
+	mux.Handle("/static/", http.StripPrefix("/static", fileServer))
+
+	log.Println("Starting server on :4000")
+	err := http.ListenAndServe(":4000", mux)
+	log.Fatal(err)
+
+}
