+++ ui/html/pages/home.tmpl.html
@@ -1,0 +1,10 @@
+{{define "title"}}
+Home{{end}}
+{{define "main"}}
+
+
+
+<h2>Latest Snippets</h2>
+<p>There's nothing to see here yet!</p>
+
+{{end}}
