+++ ui/html/partials/nav.tmpl.html
@@ -1,0 +1,5 @@
+{{define "nav"}}
+<nav>
+    <a href='/'>Home</a>
+</nav>
+{{end}}
