+++ ui/static/js/main.js
@@ -1,0 +1,8 @@
+var navLinks = document.querySelectorAll("nav a");
+for (var i = 0; i < navLinks.length; i++) {
+  var link = navLinks[i];
+  if (link.getAttribute("href") == window.location.pathname) {
+    link.classList.add("live");
+    break;
+  }
+}
